design: dir structure, database
The sum of a ssdt and dsdt table would be non-unique anyway because of the date
so save them grouped by checksum of apcidump.txt. Individual notes for the
submissions should not be necessary, rather use a per-acpidump note with
history tracking since the public should be able to edit it.